requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
coverage==7.13.4
Django==6.0.2
django-environ==0.12.0
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
Flask==2.3.2
frozenlist==1.8.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.0
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
python-telegram-bot==20.5
schedule==1.2.2
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.5
yarl==1.22.0
